pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vintage-data-lake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "blake3>=1.0.8",
    "cupy>=13.6.0",
    "cuvs-cu12>=25.10.0",
    "duckdb>=1.4.1",
    "faiss-gpu>=1.7.2",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.9",
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "numpy<2",
    "orjson>=3.11.3",
    "pandas>=2.3.3",
    "polars>=1.34.0",
    "pyarrow>=21.0.0",
    "scikit-learn>=1.7.2",
    "tiktoken>=0.12.0",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "umap-learn[plot]>=0.5.9.post2",
    "zstandard>=0.25.0",
]
